# src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API_ECOMMERCE",
    "description": "Una API para manejar todo lo relacionado a un e-commerce con sus funcionalidades basicas como crear usuarios, login, crear productos entre otras funcionalidades",
    "contact": {},
    "version": "1.0"
  },
  "servers": [
    {
      "url": "https://db-ecommerce.onrender.com",
      "variables": {}
    }
  ],
  "paths": {
    "/users": {
      "post": {
        "tags": [
          "USER"
        ],
        "summary": "crea un nuevo usuario",
        "operationId": "createnewuser",
        "parameters": [],
        "requestBody": {
          "description": "Endpoint para crear un nuevo usuario con sus datos basicos al crear el usuario se crea el carrito",
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/createnewuserrequest"
                  },
                  {
                    "example": {
                      "username": "Vinicio Chavarria",
                      "email": "vichavarria21@gmail.com",
                      "password": "12345678"
                    }
                  }
                ]
              },
              "example": {
                "username": "Vinicio Chavarria",
                "email": "vichavarria21@gmail.com",
                "password": "12345678"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      }
    },
...
    "/orders/1": {
      "put": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "Orders"
        ],
        "summary": "Completa la orden",
        "operationId": "CompletedOrder",
        "parameters": [],
        "requestBody": {
          "description": "Con este endpoint se completa la orden del usuario ",
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/CompletedOrderRequest"
                  },
                  {
                    "example": {
                      "carId": 1,
                      "email": "vichavarria21@gmail.com"
                    }
                  }
                ]
              },
              "example": {
                "carId": 1,
                "email": "vichavarria21@gmail.com"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      }
    },
...
  },
  "components": {
    "schemas": {
      "createnewuserrequest": {
        "title": "createnewuserrequest",
        "required": [
          "username",
          "email",
          "password"
        ],
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "example": {
          "username": "Vinicio Chavarria",
          "email": "vichavarria21@gmail.com",
          "password": "12345678"
        }
      },
...
      "CompletedOrderRequest": {
        "title": "CompletedOrderRequest",
        "required": [
          "carId",
          "email"
        ],
        "type": "object",
        "properties": {
          "carId": {
            "type": "integer",
            "format": "int32"
          },
          "email": {
            "type": "string"
          }
        },
        "example": {
          "carId": 1,
          "email": "vichavarria21@gmail.com"
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}